Fix v7.1.5 Transport Sweep skipped-patient accounting
