--- .github/workflows/apply-issue527-transport-sweep-skipped.yml
@@ -1,102 +1,0 @@
-name: Apply Issue 527 Transport Sweep Skipped Patients Fix
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/issue527/apply_fix.py"
-      - ".github/workflows/apply-issue527-transport-sweep-skipped.yml"
-
-permissions:
-  contents: read
-
-concurrency:
-  group: toolkit-issue-527-transport-sweep-skipped
-  cancel-in-progress: false
-
-jobs:
-  apply:
-    if: github.actor == 'Conroy1988' && github.head_ref == 'fix/issue-527-transport-sweep-skipped-patients'
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    steps:
-      - name: Check out exact Issue 527 branch
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: fix/issue-527-transport-sweep-skipped-patients
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Apply idempotent skipped-patient accounting
-        run: python3 .github/issue527/apply_fix.py
-
-      - name: Make Issue 523 contract patch-release safe
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/scripts/test_issue523_transport_sweep_progress.py')
-          text = path.read_text(encoding='utf-8')
-          old = " assert metadata and runtime and metadata.group(1)==runtime.group(1)=='7.1.4'\n"
-          new = " assert metadata and runtime and metadata.group(1)==runtime.group(1)\n version=tuple(int(part) for part in metadata.group(1).split('.'))\n assert version >= (7,1,4)\n"
-          if text.count(old) != 1:
-              raise SystemExit(f'Expected one exact v7.1.4 assertion, found {text.count(old)}')
-          path.write_text(text.replace(old, new, 1), encoding='utf-8')
-          PY
-
-      - name: Validate complete v7.1.5 candidate
-        shell: bash
-        run: |
-          set -o pipefail
-          set +e
-          {
-            python3 .github/scripts/validate_userscript.py &&
-            bash .github/scripts/run_userscript_preflight.sh --contracts &&
-            python3 .github/scripts/test_documentation_consistency.py &&
-            node --check src/MissionChief_Map_Command_Toolkit.user.js &&
-            cmp --silent dist/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.txt
-          } 2>&1 | tee issue527-validation.log
-          exit ${PIPESTATUS[0]}
-
-      - name: Upload Issue 527 validation diagnostic
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: issue527-validation-${{ github.sha }}
-          path: issue527-validation.log
-          if-no-files-found: error
-          retention-days: 14
-
-      - name: Enforce unchanged performance envelope
-        shell: bash
-        run: |
-          set -euo pipefail
-          git fetch origin main --depth=1
-          git show origin/main:src/MissionChief_Map_Command_Toolkit.user.js > /tmp/performance-base.user.js
-          python3 .github/scripts/test_performance_budget.py
-          python3 .github/scripts/check_performance_budget.py \
-            --candidate src/MissionChief_Map_Command_Toolkit.user.js \
-            --base /tmp/performance-base.user.js \
-            --policy .github/performance-budget.json \
-            --json-output /tmp/issue527-performance.json \
-            --markdown-output /tmp/issue527-performance.md
-          cat /tmp/issue527-performance.md
-
-      - name: Commit exact v7.1.5 candidate
-        env:
-          DEVELOPMENT_PR_TOKEN: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          test -n "$DEVELOPMENT_PR_TOKEN"
-          test "$GITHUB_HEAD_REF" = "fix/issue-527-transport-sweep-skipped-patients"
-          rm -f issue527-validation.log
-          git config user.name "Conroy1988"
-          git config user.email "27301455+Conroy1988@users.noreply.github.com"
-          git add -A
-          git diff --cached --quiet && { echo "No Issue 527 changes produced."; exit 1; }
-          git commit -m "Fix v7.1.5 Transport Sweep skipped-patient accounting"
-          BASIC_AUTH="$(printf 'x-access-token:%s' "$DEVELOPMENT_PR_TOKEN" | base64 -w0)"
-          git -c http.https://github.com/.extraheader="AUTHORIZATION: basic ${BASIC_AUTH}" \
-            push origin HEAD:refs/heads/fix/issue-527-transport-sweep-skipped-patients
