+++ help/index.html
@@ -1,8 +1,8 @@
 <!doctype html>
-<html lang="en-GB"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1,viewport-fit=cover"><meta name="color-scheme" content="dark"><meta name="description" content="Guide for Toolkit v7.1.4 candidate — Incident Command Wire reduced-motion autoplay recovery for The One We Knew Before."><title>MissionChief Map Command Toolkit — v7 Help Centre</title><style>
+<html lang="en-GB"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1,viewport-fit=cover"><meta name="color-scheme" content="dark"><meta name="description" content="Guide for Toolkit v7.1.5 candidate — Incident Command Wire reduced-motion autoplay recovery for The One We Knew Before."><title>MissionChief Map Command Toolkit — v7 Help Centre</title><style>
 :root{--bg:#061018;--panel:#0d2230;--line:#285066;--text:#edf8fc;--muted:#9db6c2;--cyan:#62d9ff;--gold:#ffd178}*{box-sizing:border-box}body{margin:0;background:linear-gradient(180deg,#061018,#0a1922);color:var(--text);font:15px/1.6 system-ui,-apple-system,"Segoe UI",sans-serif}header,main,footer{max-width:1180px;margin:auto;padding:22px}header{padding-top:48px}h1{font-size:clamp(38px,7vw,72px);line-height:1;margin:0}h2{margin-top:0;color:var(--cyan)}.tag{display:inline-block;padding:6px 10px;border:1px solid #6b572b;border-radius:999px;color:var(--gold)}.lead{max-width:900px;color:#c4d9e2;font-size:18px}.grid{display:grid;grid-template-columns:repeat(auto-fit,minmax(250px,1fr));gap:14px}.card,section{border:1px solid var(--line);border-radius:16px;background:rgba(13,34,48,.9);padding:20px}.card h3{margin:0 0 8px}.card p,li{color:#bfd0d8}.notice{border-color:#7f642b;background:#2c220d}.actions{display:flex;flex-wrap:wrap;gap:10px}.actions a{padding:10px 14px;border-radius:10px;background:#176f93;color:white;text-decoration:none;font-weight:700}code,kbd{background:#142d39;border:1px solid #34596a;border-radius:6px;padding:2px 6px}footer{color:var(--muted);text-align:center}@media(max-width:600px){header,main{padding:16px}header{padding-top:32px}}</style></head><body>
-<header><span class="tag">Guide for Toolkit v7.1.4 candidate</span><h1>The One We Knew Before</h1><p class="lead">Version 7 restores a clear product boundary: native MissionChief map command, fleet identity, transport support and financial intelligence without competing mission-window engines.</p><div class="actions"><a href="https://update.greasyfork.org/scripts/586018/MissionChief%20Map%20Command%20Toolkit.user.js">Install</a><a href="https://github.com/Conroy1988/missionchief-toolkit-assets/releases">Releases</a></div></header>
-<main><section class="notice"><h2>What changed in v7.1.4</h2><p>Patient Transport Sweep now keeps its mission-position counter synchronized with the immutable queue. Patient releases never increment mission progress; skips and recoverable mission errors advance once, cancellation preserves the last position and normal completion reaches total/total.</p></section>
+<header><span class="tag">Guide for Toolkit v7.1.5 candidate</span><h1>The One We Knew Before</h1><p class="lead">Version 7 restores a clear product boundary: native MissionChief map command, fleet identity, transport support and financial intelligence without competing mission-window engines.</p><div class="actions"><a href="https://update.greasyfork.org/scripts/586018/MissionChief%20Map%20Command%20Toolkit.user.js">Install</a><a href="https://github.com/Conroy1988/missionchief-toolkit-assets/releases">Releases</a></div></header>
+<main><section class="notice"><h2>What changed in v7.1.5</h2><p>Patient Transport Sweep now counts each identifiable patient vehicle skipped because no usable native Discharge patient control was available. The HUD and panel update immediately, duplicate observations cannot double-count, and mission progress remains independent.</p></section>
 <section><h2>Native Toolkit systems</h2><div class="grid">
 <div class="card"><h3>Mission Age map timers</h3><p>Compact age badges above personal missions. Press <kbd>6</kbd> to toggle them.</p></div>
 <div class="card"><h3>Mission Value</h3><p>Shows verified mission value in opened MissionChief windows without replacing native controls.</p></div>
@@ -20,4 +20,4 @@
 <section><h2>Seven complete interface systems</h2><p>Map Command, Cyberpunk, Fallout 4, Umbrella, Factorio, 007 Intelligence and Hyrule Command share the same retained native feature set.</p></section>
 <section><h2>Device support</h2><p>Desktop, Tablet/iPad and iOS Mobile Mode remain supported with safe-area handling, 44px touch targets, orientation recovery and responsive panels.</p></section>
 <section><h2>Privacy</h2><p>Most settings stay in the browser. Financial reporting uses only the saved Discord webhook supplied by the user; exported settings containing it should be treated as private.</p></section></main>
-<footer>MissionChief Map Command Toolkit · v7.1.4 candidate · The One We Knew Before</footer></body></html>
+<footer>MissionChief Map Command Toolkit · v7.1.5 candidate · The One We Knew Before</footer></body></html>
